refactor: move resource helpers to server side
